backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0